Add files via upload
Introduction

DIVA

DIVA (Damn insecure and vulnerable App) is an App intentionally designed to be insecure. It is Android version of Diva. It would be a nice way to start the year by contributing something to the security community. The aim of the App is to teach developers/QA/security professionals, flaws that are generally present in the Apps due poor or insecure coding practices. If anyone reading this, you want to either learn App pen testing or secure coding and I sincerely hope that DIVA solves your purpose.

Who can use Diva

The idea originated, from a developer’s perspective. The Android security training for developers becomes slightly boring with lot of theory and not much hands-on. Diva gamifies secure development learning. It is an excellent learning tool for aspiring Android penetration testers and security professionals as it gives an insight into app vulnerabilities including the source code. To sum it up:
•	Android App developers
•	Android Penetration testers
•	Security professionals
•	Students

Vulnerabilities include:

•	Insecure Logging
•	Hardcoding Issues – Part 1
•	Insecure Data Storage – Part 1
•	Insecure Data Storage – Part 2
•	Insecure Data Storage – Part 3
•	Insecure Data Storage – Part 4
•	Input Validation Issues – Part 1
•	Input Validation Issues – Part 2
•	Access Control Issues – Part 1
•	Access Control Issues – Part 2
•	Access Control Issues – Part 3
•	Hardcoding Issues – Part 2
•	Input Validation Issues – Part 3